moved local notifications code to a new utilities ts file
